feat: Add delay method for function delay
Add a new method, delay, to the DurationExtensions class in num_extensions.dart. This method allows for delaying the execution of a function for a specified duration. The delay method utilizes the Future.delayed function to pause the execution of the callback function for the given duration. This feature enhances the flexibility and usability of the DurationExtensions class, enabling developers to easily introduce delays in their code.