refactor: Update isValid method to use customValidator returning a list of booleans
Update the isValid method in global.dart to use a customValidator parameter that returns a list of booleans. This change improves the flexibility of the isValid method, allowing developers to define multiple validation functions and determine the validity of an object based on the results of those functions. This refactor enhances the overall quality of the codebase and improves the extensibility of the isValid method.