refactor: Update SelfMap class to use Iterable instead of List for items parameter
Update the SelfMap class in self_map.dart to use an Iterable instead of a List for the items parameter. This change improves the flexibility of the SelfMap class, allowing developers to use any iterable collection as the input for creating the map. This refactor enhances the overall quality and usability of the codebase.