refactor: Update `ScaffoldBuilder` to support custom title color
Update the `ScaffoldBuilder` class in `scaffold_builder.dart` to support a custom title color. This change adds a new `titleColor` property to the `ScaffoldBuilder` constructor and sets the `foregroundColor` of the app bar to the specified title color. This refactor enhances the flexibility and customization options of the `ScaffoldBuilder` widget.