refactor: Improve `isIn` method in `general_extensions.dart`
Improve the `isIn` method in `general_extensions.dart` to support checking if the current object is present in a list, map, or string. This enhancement adds support for checking if the object is present in a map by checking its values, and in a string by checking if it is one of the words in the string. This refactor improves the functionality and flexibility of the `isIn` method, allowing for more versatile object presence checks.