refactor: Update `KeyedJsonTable` getPair and getPairs methods
Refactor the `getPair` and `getPairs` methods in the `KeyedJsonTable` class to use a single function parameter instead of separate keyFunc and valueFunc parameters. This simplifies the method signatures and improves code readability. The new function parameter should accept a `JsonRow` and return a pair of values of types `T1` and `T2`. This refactor enhances the maintainability and flexibility of the `KeyedJsonTable` class.