refactor: Update `nfe.dart` to include randomItem and randomWord methods
Add the `randomItem` method to the `List` extension in `list_extensions.dart` to return a random item from the list. Also, include the `randomWord` function in `global.dart` to generate a random string of specified length. These changes enhance the functionality of the code by providing convenient methods for working with random data.