refactor: Update `global.dart` and `brasil.dart` files
The `global.dart` file was modified to update the `random` variable to `_random`. This change improves the encapsulation of the random number generator.

The `brasil.dart` file was modified to refactor the `gerarCNPJFake` and `gerarCPFFake` methods. The methods now use the `randomInt` function instead of `random.nextInt` to generate random numbers. This change ensures consistent and reliable fake CNPJ and CPF generation.